// FrontEnd/src/app/Enums/security-question.enum.ts
export enum SecurityQuestion {
    WhatIsYourFavoriteColor = 0,
    WhatCityWereYouBornIn = 1,
    WhatIsYourMotherMaidenName = 2,
    WhatIsTheNameOfYourFirstPet = 3,
    WhatIsYourFavoriteBook = 4,
    WhatWasTheNameOfYourElementarySchool = 5,
    WhatIsYourFavoriteChildhoodMovie = 6,
    WhatIsTheNameOfYourBestChildhoodFriend = 7,
    WhatWasTheMakeOfYourFirstCar = 8,
    WhatIsTheNameOfTheStreetYouGrewUpOn = 9
  }
  